+++ drivers/modem/CMakeLists.txt
@@ -4,8 +4,8 @@ zephyr_library()
 
 zephyr_library_include_directories(${ZEPHYR_BASE}/subsys/net/ip)
 zephyr_library_include_directories(${ZEPHYR_BASE}/subsys/net/lib/sockets)
-add_subdirectory(simcom)
-add_subdirectory_ifdef(CONFIG_MODEM_HL78XX hl78xx)
+
+add_subdirectory(vendor_standalone)
 add_subdirectory_ifdef(CONFIG_MODEM_CELLULAR vendor_modem_cellular)
 
 # zephyr-keep-sorted-start
@@ -14,14 +14,9 @@ zephyr_library_sources_ifdef(CONFIG_MODEM_AT_USER_PIPE modem_at_user_pipe.c)
 zephyr_library_sources_ifdef(CONFIG_MODEM_CELLULAR modem_cellular.c)
 zephyr_library_sources_ifdef(CONFIG_MODEM_CMD_HANDLER modem_cmd_handler.c)
 zephyr_library_sources_ifdef(CONFIG_MODEM_CONTEXT modem_context.c)
-zephyr_library_sources_ifdef(CONFIG_MODEM_HL7800 hl7800.c)
 zephyr_library_sources_ifdef(CONFIG_MODEM_IFACE_UART_ASYNC modem_iface_uart_async.c)
 zephyr_library_sources_ifdef(CONFIG_MODEM_IFACE_UART_INTERRUPT modem_iface_uart_interrupt.c)
-zephyr_library_sources_ifdef(CONFIG_MODEM_QUECTEL_BG9X quectel-bg9x.c)
 zephyr_library_sources_ifdef(CONFIG_MODEM_RECEIVER modem_receiver.c)
 zephyr_library_sources_ifdef(CONFIG_MODEM_SHELL modem_shell.c)
 zephyr_library_sources_ifdef(CONFIG_MODEM_SOCKET modem_socket.c)
-zephyr_library_sources_ifdef(CONFIG_MODEM_ST87MXX st87mxx.c)
-zephyr_library_sources_ifdef(CONFIG_MODEM_UBLOX_SARA ublox-sara-r4.c)
-zephyr_library_sources_ifdef(CONFIG_MODEM_WNCM14A2A wncm14a2a.c)
 # zephyr-keep-sorted-stop
