+++ drivers/modem/Kconfig
@@ -191,13 +191,7 @@ config MODEM_CELL_INFO
 # zephyr-keep-sorted-start
 source "drivers/modem/Kconfig.at_shell"
 source "drivers/modem/Kconfig.cellular"
-source "drivers/modem/Kconfig.hl7800"
-source "drivers/modem/Kconfig.quectel-bg9x"
-source "drivers/modem/Kconfig.st87mxx"
-source "drivers/modem/Kconfig.ublox-sara-r4"
-source "drivers/modem/Kconfig.wncm14a2a"
-source "drivers/modem/hl78xx/Kconfig.hl78xx"
-source "drivers/modem/simcom/Kconfig"
+source "drivers/modem/vendor_standalone/Kconfig"
 # zephyr-keep-sorted-stop
 
 endif # MODEM
