+++ drivers/modem/vendor_standalone/hl78xx/CMakeLists.txt
