+++ drivers/modem/vendor_standalone/hl78xx/gnss_linker.ld
