+++ drivers/modem/vendor_standalone/hl78xx/hl78xx_evt_monitor/hl78xx_evt_monitor.ld
