+++ drivers/modem/vendor_standalone/hl78xx/hl78xx_gnss.c
