+++ drivers/modem/vendor_standalone/hl78xx/hl78xx_hl7800.c
