+++ drivers/modem/vendor_standalone/hl78xx/hl78xx_hl7812.c
