+++ drivers/modem/vendor_standalone/simcom/CMakeLists.txt
