+++ drivers/modem/vendor_standalone/simcom/sim7080/sim7080_utils.c
