+++ drivers/modem/vendor_standalone/wncm14a2a.c
@@ -37,7 +37,7 @@ LOG_MODULE_REGISTER(LOG_DOMAIN);
 #include "udp_internal.h"
 #endif
 
-#include "modem_receiver.h"
+#include "../modem_receiver.h"
 
 /* Uncomment the #define below to enable a hexdump of all incoming
  * data from the modem receiver
